Services are running on 4 Containers on localhost. Added Kubernetes Manifest Pods not working. MySQL Databases working. Terraform - Infrastructure runs on EXOSCALE correctly
