+++ k8s/mysql-customer.yaml
@@ -1,0 +1,59 @@
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: mysql-customer
+spec:
+  replicas: 1
+  selector:
+    matchLabels:
+      app: mysql-customer
+  template:
+    metadata:
+      labels:
+        app: mysql-customer
+    spec:
+      containers:
+      - name: mysql-customer
+        image: mysql:5.7
+        imagePullPolicy: IfNotPresent
+        env:
+        - name: MYSQL_ROOT_PASSWORD
+          value: "123456789"  # Ersetzen Sie diesen Wert durch ein sicheres Passwort
+        - name: MYSQL_DATABASE
+          value: "customer_db"   # Der Name der Datenbank
+        - name: MYSQL_USER
+          value: "customer_user" # Der Name des neuen Benutzers
+        - name: MYSQL_PASSWORD
+          value: "Customer1" # Ersetzen Sie diesen Wert durch ein sicheres Passwort für den neuen Benutzer
+        ports:
+        - containerPort: 3306
+        volumeMounts:
+        - name: mysql-customer-pv
+          mountPath: /var/lib/mysql
+      volumes:
+      - name: mysql-customer-pv
+        persistentVolumeClaim:
+          claimName: mysql-customer-pvc
+---
+apiVersion: v1
+kind: Service
+metadata:
+  name: mysql-customer
+spec:
+  selector:
+    app: mysql-customer
+  ports:
+  - protocol: TCP
+    port: 3306
+    targetPort: 3306
+---
+apiVersion: v1
+kind: PersistentVolumeClaim
+metadata:
+  name: mysql-customer-pvc
+spec:
+  accessModes:
+  - ReadWriteOnce
+  resources:
+    requests:
+      storage: 10Gi
