+++ k8s/mysql-room.yaml
@@ -1,0 +1,59 @@
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: mysql-room
+spec:
+  replicas: 1
+  selector:
+    matchLabels:
+      app: mysql-room
+  template:
+    metadata:
+      labels:
+        app: mysql-room
+    spec:
+      containers:
+      - name: mysql-room
+        image: mysql:5.7
+        imagePullPolicy: IfNotPresent
+        env:
+        - name: MYSQL_ROOT_PASSWORD
+          value: "123456789"
+        - name: MYSQL_DATABASE
+          value: "room_db"
+        - name: MYSQL_USER
+          value: "room_user"
+        - name: MYSQL_PASSWORD
+          value: "Room1"
+        ports:
+        - containerPort: 3306
+        volumeMounts:
+        - name: mysql-room-pv
+          mountPath: /var/lib/mysql
+      volumes:
+      - name: mysql-room-pv
+        persistentVolumeClaim:
+          claimName: mysql-room-pvc
+---
+apiVersion: v1
+kind: Service
+metadata:
+  name: mysql-room
+spec:
+  selector:
+    app: mysql-room
+  ports:
+  - protocol: TCP
+    port: 3306
+    targetPort: 3306
+---
+apiVersion: v1
+kind: PersistentVolumeClaim
+metadata:
+  name: mysql-room-pvc
+spec:
+  accessModes:
+  - ReadWriteOnce
+  resources:
+    requests:
+      storage: 10Gi
