+++ terraform/provider.yaml
@@ -1,0 +1,6 @@
+apiVersion: pkg.crossplane.io/v1
+kind: Provider
+metadata:
+   name: provider-github
+spec:
+   package: xpkg.upbound.io/coopnorge/provider-github:v0.10.0
