+++ terraform/service.yaml
@@ -1,0 +1,11 @@
+apiVersion: v1
+kind: Service
+metadata:
+  name: podtato-head
+spec:
+    selector:
+        app: podtato-head
+    ports:
+    - protocol: TCP
+      port: 8080
+      targetPort: 8080
