+++ terraform/terraform.tfstate
@@ -1,0 +1,9 @@
+{
+  "version": 4,
+  "terraform_version": "1.6.6",
+  "serial": 107,
+  "lineage": "942835bd-6f02-53bc-b625-57fc600b8259",
+  "outputs": {},
+  "resources": [],
+  "check_results": null
+}
