+++ terraform/vars.tf
@@ -1,0 +1,11 @@
+variable "exo_secret" {
+  type      = string
+  sensitive = true
+  default   = "ll1mYK9P5O3xx52P6ftg3vnrFRyNPkdW4PV3oB_NLo8"
+}
+
+variable "exo_key" {
+  type      = string
+  sensitive = true
+  default   = "EXO5f27ae1ba3685a0d89d9ae58"
+}
